requirements.txt
flask
flask-cors
gunicorn
python-dotenv
transformers==4.32.0
torch==2.0.0
sentence-transformers
langchain-community
langchain-text-splitters
